Final remediation: production hardening, bug fixes, and policy alignment
- Enforce fail-closed security with ProductionRuntimeGuard in 22 native managers.
- Correct BdkManager descriptor interpolation and align EVM/SilentPayment signatures.
- Resolve Gitleaks findings in tests and purge generated artifacts.
- Implement SDK Ownership and Control-Plane policies (CON-1178, CON-773).
- Synchronize README, CODEOWNERS, and documentation for v1.9.2.
- Verified with 250+ passing tests.

+++ android/core-bitcoin/src/main/kotlin/com/conxius/wallet/bitcoin/EvmManager.kt
@@ -11,7 +11,7 @@ class EvmManager {
     /**
      * Signs an EVM transaction (Legacy, EIP-1559, or EIP-2930).
      */
-    fun signTransaction(txPayload: ByteArray, chainId: Int): String {
+    fun signTransaction(txPayload: ByteArray, chainId: Long): String {
         Log.d(TAG, "Signing EVM Transaction for chain: $chainId")
         return ProductionRuntimeGuard.failClosed(
             "EVM transaction signing",
